+++ src/commands/clean.ts
@@ -16,6 +16,7 @@
 
 import { Command } from '@oclif/command';
 import { BootstrapUtils } from '../service';
+import { CommandUtils } from '../service/CommandUtils';
 
 export default class Clean extends Command {
     static description = 'It removes the target folder deleting the generated configuration and data';
@@ -23,8 +24,8 @@ export default class Clean extends Command {
     static examples = [`$ symbol-bootstrap clean`];
 
     static flags = {
-        help: BootstrapUtils.helpFlag,
-        target: BootstrapUtils.targetFlag,
+        help: CommandUtils.helpFlag,
+        target: CommandUtils.targetFlag,
     };
 
     public async run(): Promise<void> {