+++ src/commands/compose.ts
@@ -16,6 +16,7 @@
 
 import { Command, flags } from '@oclif/command';
 import { BootstrapService, BootstrapUtils, ComposeService } from '../service';
+import { CommandUtils } from '../service/CommandUtils';
 
 export default class Compose extends Command {
     static description = 'It generates the `docker-compose.yml` file from the configured network.';
@@ -23,9 +24,10 @@ export default class Compose extends Command {
     static examples = [`$ symbol-bootstrap compose`];
 
     static flags = {
-        help: BootstrapUtils.helpFlag,
-        target: BootstrapUtils.targetFlag,
-        password: BootstrapUtils.passwordFlag,
+        help: CommandUtils.helpFlag,
+        target: CommandUtils.targetFlag,
+        password: CommandUtils.passwordFlag,
+        noPassword: CommandUtils.noPasswordFlag,
         upgrade: flags.boolean({
             description: 'It regenerates the docker compose and utility files from the <target>/docker folder',
             default: ComposeService.defaultParams.upgrade,
@@ -40,6 +42,7 @@ export default class Compose extends Command {
     public async run(): Promise<void> {
         const { flags } = this.parse(Compose);
         BootstrapUtils.showBanner();
+        flags.password = await CommandUtils.resolvePassword(flags.password, flags.noPassword, CommandUtils.passwordPromptDefaultMessage);
         await new BootstrapService(this.config.root).compose(flags);
     }
 }