+++ src/commands/config.ts
@@ -16,6 +16,7 @@
 
 import { Command, flags } from '@oclif/command';
 import { BootstrapService, BootstrapUtils, ConfigService, Preset } from '../service';
+import { CommandUtils } from '../service/CommandUtils';
 
 export default class Config extends Command {
     static description = 'Command used to set up the configuration files and the nemesis block for the current network';
@@ -23,9 +24,10 @@ export default class Config extends Command {
     static examples = [`$ symbol-bootstrap config -p bootstrap`];
 
     static flags = {
-        help: BootstrapUtils.helpFlag,
-        target: BootstrapUtils.targetFlag,
-        password: BootstrapUtils.passwordFlag,
+        help: CommandUtils.helpFlag,
+        target: CommandUtils.targetFlag,
+        password: CommandUtils.passwordFlag,
+        noPassword: CommandUtils.noPasswordFlag,
         preset: flags.enum({
             char: 'p',
             description: 'the network preset',
@@ -74,6 +76,7 @@ export default class Config extends Command {
     public async run(): Promise<void> {
         const { flags } = this.parse(Config);
         BootstrapUtils.showBanner();
+        flags.password = await CommandUtils.resolvePassword(flags.password, flags.noPassword, CommandUtils.passwordPromptDefaultMessage);
         await new BootstrapService(this.config.root).config(flags);
     }
 }