+++ src/commands/enrolRewardProgram.ts
@@ -17,6 +17,7 @@
 import { Command } from '@oclif/command';
 import { BootstrapService, BootstrapUtils } from '../service';
 import { AnnounceService } from '../service/AnnounceService';
+import { CommandUtils } from '../service/CommandUtils';
 
 export default class EnrolRewardProgram extends Command {
     static description = `It enrols the nodes in the rewards program by announcing the enrol transaction to the registration address.  You can also use this command to update the program registration when you change the node public key or server host.
@@ -26,8 +27,8 @@ Currently, the only program that can be enrolled post-launch is 'SuperNode'.`;
     static examples = [`$ symbol-bootstrap enrolRewardProgram`];
 
     static flags = {
-        help: BootstrapUtils.helpFlag,
-        target: BootstrapUtils.targetFlag,
+        help: CommandUtils.helpFlag,
+        target: CommandUtils.targetFlag,
         ...AnnounceService.flags,
     };